+++ examples/user-guide/09-topsystem.py
@@ -418,9 +418,10 @@ riv_stage = enforce_dim_order(riv_stage)
 riv_stage
 
 # %%
+#
 # We can do the same for the river bottom and construct a river package. Note
 # that we use the previously distributed conductance which we assigned to the
-# variable ``riv_conducance``.
+# variable ``riv_conductance``.
 
 riv_bottom = planar_river["bottom"].where(riv_allocated)
 riv_bottom = enforce_dim_order(riv_bottom)
@@ -434,5 +435,50 @@ riv = imod.mf6.River(
 
 riv
 
+# %%
+#
+# Reallocate package
+# ------------------
+#
+# The river package has a :meth:`imod.mf6.River.reallocate` method which can be
+# used to reallocate the river package to a new model layer schematization. This
+# is convenient when you already have an existing model, but want to apply a
+# different allocation or distribution option to its river package. This saves
+# you from unpacking the right variables from the the DIS/DISV and NPF package,
+# aggregating the package data over layers and allocating and distributing the
+# conductance again. There are equivalent methods for the
+# :meth:`imod.mf6.Drainage.reallocate`,
+# :meth:`imod.mf6.GeneralHeadBoundary.reallocate`, and
+# :meth:`imod.mf6.Recharge.reallocate`.
+
+dis = imod.mf6.StructuredDiscretization(
+    top=layer_model["top"].sel(layer=1),
+    bottom=layer_model["bottom"],
+    idomain=layer_model["idomain"].astype(int),
+)
+npf = imod.mf6.NodePropertyFlow(icelltype=0, k=layer_model["k"])
+
+riv_reallocated = riv.reallocate(
+    dis, npf, allocation_option=ALLOCATION_OPTION.stage_to_riv_bot
+)
+
+riv_reallocated
 
 # %%
+#
+# The default allocation option and distribution option for the are set in the
+# :class:`imod.prepare.SimulationAllocationOptions` and
+# :class:`imod.prepare.SimulationDistributingOptions`. Let's print the default
+# options to see what they are. First let's start with the allocation options:
+
+from dataclasses import asdict
+
+from imod.prepare import SimulationAllocationOptions, SimulationDistributingOptions
+
+print(asdict(SimulationAllocationOptions()))
+
+# %%
+#
+# Now let's print the distribution options:
+
+print(asdict(SimulationDistributingOptions()))