+++ imod/mf6/simulation.py
@@ -23,9 +23,9 @@ import imod.mf6.exchangebase
 from imod.common.interfaces.imodel import IModel
 from imod.common.interfaces.isimulation import ISimulation
 from imod.common.statusinfo import NestedStatusInfo
+from imod.common.utilities.dataclass_type import DataclassType
 from imod.common.utilities.mask import _mask_all_models
 from imod.common.utilities.regrid import _regrid_like
-from imod.common.utilities.regrid_method_type import RegridMethodType
 from imod.common.utilities.version import (
     get_version,
     log_versions,
@@ -282,7 +282,7 @@ class Modflow6Simulation(collections.UserDict, ISimulation):
         ----------
         directory: str, pathlib.Path
             Directory to write Modflow 6 simulation to.
-        use_binary: ({True, False}, optional)
+        binary: ({True, False}, optional)
             Whether to write time-dependent input for stress packages as binary
             files, which are smaller in size, or more human-readable text files.
         validate: ({True, False}, optional)
@@ -1412,7 +1412,7 @@ class Modflow6Simulation(collections.UserDict, ISimulation):
         times: list[datetime],
         allocation_options: Optional[SimulationAllocationOptions] = None,
         distributing_options: Optional[SimulationDistributingOptions] = None,
-        regridder_types: Optional[dict[str, RegridMethodType]] = None,
+        regridder_types: Optional[dict[str, DataclassType]] = None,
     ) -> "Modflow6Simulation":
         """
         Imports a GroundwaterFlowModel (GWF) from the data in an iMOD5 project