+++ imod/msw/vegetation.py
@@ -4,7 +4,7 @@ import numpy as np
 import xarray as xr
 
 from imod.common.interfaces.iregridpackage import IRegridPackage
-from imod.common.utilities.regrid_method_type import EmptyRegridMethod, RegridMethodType
+from imod.common.utilities.dataclass_type import DataclassType, EmptyRegridMethod
 from imod.msw.fixed_format import VariableMetaData
 from imod.msw.pkgbase import MetaSwapPackage
 
@@ -56,7 +56,7 @@ class AnnualCropFactors(MetaSwapPackage, IRegridPackage):
         "ponding_factor": VariableMetaData(8, 0.01, 10.0, float),
     }
 
-    _regrid_method: RegridMethodType = EmptyRegridMethod()
+    _regrid_method: DataclassType = EmptyRegridMethod()
 
     def __init__(
         self,